better tests

+++ src/main/java/main/batch/steps/step3/processors/Step3ItemProcessor.java
@@ -4,6 +4,8 @@ import main.databases.mysql.domain.PersonMySQL;
 import main.databases.postgresql.domain.PersonPostgreSQL;
 import org.springframework.batch.item.ItemProcessor;
 
+import java.math.BigDecimal;
+
 public class Step3ItemProcessor implements ItemProcessor<PersonMySQL, PersonPostgreSQL> {
 
     @Override
@@ -11,7 +13,7 @@ public class Step3ItemProcessor implements ItemProcessor<PersonMySQL, PersonPost
         PersonPostgreSQL outputDataModel = new PersonPostgreSQL();
         outputDataModel.setId(inputDataModel.getId());
         outputDataModel.setName(inputDataModel.getName());
-        outputDataModel.setSalary(inputDataModel.getSalary());
+        outputDataModel.setSalary(BigDecimal.valueOf(inputDataModel.getSalary()));
         outputDataModel.setAge(inputDataModel.getAge());
         return outputDataModel;
     }
