+++ src/main/resources/db/migration/mysql/V1__Initial.sql
@@ -1,1 +1,1 @@
-CREATE TABLE person_mysql (ID INT PRIMARY KEY, name VARCHAR(50), salary DECIMAL(5,2), age int);
+CREATE TABLE person_mysql (ID INT PRIMARY KEY, name VARCHAR(50), salary DOUBLE, age int);
